Solutions

+++ exercises/02_basic_calculator/01_integers/src/lib.rs
@@ -1,6 +1,7 @@
 fn compute(a: u32, b: u32) -> u32 {
-    // TODO: change the line below to fix the compiler error and make the tests pass.
-    let multiplier: u8 = 4;
+    let multiplier: u32 = 4;
+    // ^ You could also omit `: u32` entirely.
+    // The compiler will infer the correct type based on the usage below.
     a + b * multiplier
 }
